third_party/WebKit/LayoutTests/fast/dom/location-hash-usvstring.html

<!DOCTYPE HTML>
<title>location.hash test for USVString</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<body>
<script>
test(function() {
    // USVString is equivalent to DOMString except for not allowing unpaired surrogate codepoints.
    // U+D800 to U+DFFF code point values are reserved as high and low surrogates.
    // Unpaired surrogate codepoints present in USVString are replaced with U+FFFD.
    location.hash = '\uD999';
    strActual = location.hash;
    location.hash = '\uFFFD';
    strReplace = location.hash;
    assert_equals(strActual, strReplace);
}, "unpaired surrogate codepoint should be replaced with U+FFFD");
</script>
</body>
